Revert "GraphQL API playground (#1123)"
This reverts commit d04b7f058a94ecb0201575e4125ef6236acbbb4a.
