+++ package.json
@@ -42,10 +42,8 @@
     "create-react-class": "15.6.3",
     "csurf": "1.10.0",
     "express": "4.17.1",
-    "graphql": "^14.3.1",
-    "graphql-playground-react": "^1.7.20",
     "history": "4.9.0",
-    "immutable": "^4.0.0-rc.9",
+    "immutable": "3.8.1",
     "immutable-devtools": "0.1.3",
     "js-beautify": "1.10.0",
     "json-file-plus": "3.2.0",
@@ -61,7 +59,6 @@
     "react-dnd-html5-backend": "8.0.3",
     "react-dom": "16.8.6",
     "react-helmet": "5.2.1",
-    "react-redux": "^5.1.1",
     "react-router": "5.0.1",
     "react-router-dom": "5.0.1"
   },
