+++ src/dashboard/index.js
@@ -13,7 +13,6 @@ import Dashboard       from './Dashboard';
 import '@babel/polyfill';
 
 require('stylesheets/fonts.scss');
-require('graphql-playground-react/build/static/css/index.css')
 installDevTools(Immutable);
 
 var path = window.PARSE_DASHBOARD_PATH || '/';
