+++ src/lib/ParseApp.js
@@ -42,8 +42,7 @@ export default class ParseApp {
     primaryBackgroundColor,
     secondaryBackgroundColor,
     supportedPushLocales,
-    preventSchemaEdits,
-    graphQLServerURL
+    preventSchemaEdits
   }) {
     this.name = appName;
     this.createdAt = created_at ? new Date(created_at) : new Date();
@@ -68,7 +67,6 @@ export default class ParseApp {
     this.secondaryBackgroundColor=secondaryBackgroundColor;
     this.supportedPushLocales = supportedPushLocales ? supportedPushLocales : [];
     this.preventSchemaEdits = preventSchemaEdits || false;
-    this.graphQLServerURL = graphQLServerURL;
 
     if(!supportedPushLocales) {
       console.warn(`Missing push locales for '` + appName + `', see this link for details on setting localizations up. https://github.com/parse-community/parse-dashboard#configuring-localized-push-notifications`);
